+++ Dockerfile
@@ -18,8 +18,6 @@ RUN pip install -r requirements.txt
 
 # Copy the current directory contents into the container at /app
 COPY . /app/
-# Make port 5000 available to the world outside this container
-EXPOSE 5000
 
 # Run app.py when the container launches
-CMD ["python", "server.py"]
+CMD ["python", "solweig-generator.py"]
