+++ docker-compose.yml
@@ -1,0 +1,17 @@
+version: '3'
+
+services:
+  solweig-jobs:
+    build: .
+    volumes:
+      - data-volume:/app/output
+  flask-app:
+    build: ./server
+    ports:
+      - "5000:5000"
+    volumes:
+      - data-volume:/app/output
+  
+volumes:
+  data-volume:
+
