+++ server/requirements.txt
@@ -1,0 +1,77 @@
+GDAL==3.6.4
+affine==2.4.0
+attrs==23.1.0
+blinker==1.7.0
+branca==0.6.0
+Brotli==1.1.0
+build==1.0.3
+certifi==2023.11.17
+charset-normalizer==3.3.1
+click==8.1.7
+click-plugins==1.1.1
+cligj==0.7.2
+colorama==0.4.6
+contourpy==1.1.1
+cycler==0.12.1
+docopt==0.6.2
+fiona==1.9.5
+Flask==3.0.0
+Flask-Cors==4.0.0
+Flask-APScheduler==1.13.1
+folium==0.14.0
+fonttools==4.43.1
+geojson==3.1.0
+geopandas==0.13.2
+idna==3.4
+importlib-metadata==6.8.0
+importlib-resources==6.1.0
+itsdangerous==2.1.2
+Jinja2==3.1.2
+joblib==1.3.2
+kiwisolver==1.4.5
+mapclassify==2.5.0
+MarkupSafe==2.1.3
+matplotlib==3.7.3
+munch==4.0.0
+munkres==1.1.4
+networkx==2.8.8
+numpy==1.24.4
+osmnx==1.7.0
+packaging==23.2
+pandas==2.0.3
+Pillow==10.0.0
+pip==23.3.1
+pip-tools==7.3.0
+pipreqs==0.4.13
+platformdirs==3.11.0
+pooch==1.8.0
+pyparsing==3.1.1
+pyproj==3.5.0
+pyproject_hooks==1.0.0
+PySocks==1.7.1
+pysolar==0.11
+python-dateutil==2.8.2
+python-dotenv==1.0.0
+pytz==2023.3.post1
+rasterio==1.3.7
+requests==2.31.0
+Rtree==1.1.0
+scikit-learn==1.3.2
+scipy==1.10.1
+setuptools==68.2.2
+shapely==2.0.1
+simplekml==1.3.6
+six==1.16.0
+snuggs==1.4.7
+threadpoolctl==3.2.0
+tomli==2.0.1
+typing_extensions==4.8.0
+tzdata==2023.3
+unicodedata2==15.1.0
+urllib3==2.0.7
+Werkzeug==3.0.1
+wheel==0.41.2
+win-inet-pton==1.1.0
+xyzservices==2023.10.0
+yarg==0.1.9
+zipp==3.17.0
